pyproject.toml
[project]
name = "namereplacer"
version = "0.1.0"
description = ""
authors = [
    {name = "marva-sequioaat",email = "marva@sequoiaat.com"}
]
packages = [{include = "namereplacer"}]
include = [
   "namereplacer/fetchers/mapping.csv",
]

readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "requests-mock (>=1.12.1,<2.0.0)",
  "pytest (>=8.3.4,<9.0.0)"  
]

[project.scripts]
namereplacer = "namereplacer.__main__:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
